ch9 done
Aside from the usual replacements, I should point out the \hbox
containing "(-37.5\%)" near the start. Turns out the engine can break
lines after the hyphen! I tested and it seems it won't do this if a
minus sign is used, so that's one good argument for making *those*
replacements later.