fix ch12; retroactive replacements for IV and CP
only occurrences of IV that have subscripts were replaced. as for CP,
replaced everywhere.
  note the \mbox approach for shorthands does fail in inline fractions.
in section 12.3, \frac{P\cdot\Eff{A}}{\Eff{D}} would put a (correctly)
small P and a normal-sized Eff_A and Eff_D, which looks stupid; so had
to do them manually. i think it's ok given that this only happens once
here and once in ch13, but it'd be nice to find a better approach; may
look into it another time